centralize some validations
